Cleanup specification module
